Asynchronous gerrit review commands.
Added a thread-pool and a queue to handle assembling and sending of
review commands on a separate thread so that the ToGerritRunListener
doesn't clog the builds from finishing in a timely fashion.
This required a bit of refactoring that also finally moved most ssh
related code into the gerrit-events module.