Removed connecting to gerrit on startup
As default, the url to gerrit is now blank. Unless an url is entered,
plugin will not connect to gerrit on startup. If the url is blank, a warning
will be shown reminding the user to enter an url. If an url is entered and
a connection still cannot be made, an error is shown.

Change-Id: Ibbc5e279234758be97145d9a2e1abf40409cbde1