Use event as key in BuildMemory.
Removes PatchSetKey class which was used as the key to a build memory.
Replaces it with simply using the event as they key (a custom comparator
is used to ensure that the java.lang.Object hashcode is used as the key).

Also add the hash of the event as an extra build parameter.  This allows
otherwise duplicate builds to be scheduled.

Together, these changes ensure that every Gerrit event triggers a full
complement of the jobs that are listening for it, all of them are
scheduled, and all of them report back to Gerrit as a unit, even if
multiple events of the same type and for the same change arrive.