Follow up fix: hash code shouldn't be a map key.
This is a follow-up fix to the previous commit, and one that addresses
the real underlying problem of using hash codes as the key.

Since add/remove of the listeners are done by passing in the instance,
the correct thing to do is to use the listener instance itself as the
identity, hence the use of Set, not Map.

In addition, I made the code to use CopyOnWriteArraySet so that we don't
have to maintain the code of copying around a set every time we change
it. It is also probably run faster, as we iterate over them more often
than change listeners.