Fixed circularity bug of CommentAdded.
An issue could occur when the Gerrit trigger
for a project is set to trigger on the same
CommentAdded verdict category and value as
the one that is set by the Gerrit trigger,
causing an infinite loop of builds. This change
fixes that bug.

Also fixed an issue with triggers for projects
that are set up to trigger on several verdict
categories and/or values. Previously, we stopped
at the first category/value we found, now we
iterate through them all.

Change-Id: I02a2cb29e795b7787b2902e8dc9dd9ff9e79b265