Throw IOException insteadof IllegalState
It seems to happen more often than expected that the connection
only gets half done. The main loop in the handler is designed to
exit on runtime exceptions so a retry will not be performed.
But with a checked exception it should just go back and retry the
connection instead.

Change-Id: Ib80893ee9cc74a7baa7ca92d5dcbf5677e7b4dad