Move GerritEventLifecycle and GerritEventLifecycleListener to gerrit-trigger module
These 2 classes are related to the lifecycle of and event in Jenkins so they do not
belong in the gerrit-events module.

Converted the GerritEventLifecycle to an interface that so we could created adaptor
classes in the gerrit-trigger to add lifecycle to any type of event. So far, only
ManualPatchsetCreated requires lifecycle so it is now implementing this interface;
no need for adaptor since ManualPatchset is defined in gerrit-trigger. Also created
a sub class of GerritHandler in gerrit-trigger that is able to fire the lifecycle
events.