Add RepositoryModifiedEvent interface
RepositoryModifiedEvent represents any event raised by a ref created/changed in
the git repository (PatchsetCreated, DraftCreated and RefUpdated).

An example of usage is the gerrit-trigger: to implement wait for
replication feature, when a GerritEvent is interesting, it must wait
for replication to be complete before starting the build but only for
events that created/changed a ref in the git repository.

Change-Id: Icaec80452e4753ee80b38084978953578b488514