Add Change-Id to comparison target
Now some members in Change class is used for object comparison. Change
number is one of them.

It has site scope. But if client supports multisite gerrit, it has
possibility to conflict with ones of other sites. It means that new
member which has smaller scope should be needed.

This patch provides Change-Id to object comparison logic.