fix: JENKINS-44568, JENKINS-44414
When a message is received that is bigger than the internal buffer, the
internal buffer position was wrongly set to the buffer capacity.
But we also can't just delete the buffer is no newline was found. So we
append what we've read in a Stringbuilder and append until we encounter a
newline.