fix a potential corruption of data
in case the data is bigger than the buffer and the data contains one or
several blanks that end up at the beginning of the buffer those blanks
would have been removed. So we have to trim only right before we return
the line.
Avoid possible performance bottle neck in case of a single event that
fills the buffer multiple times