Filter out duplicate LineComments
Exact duplicate LineComments don't make sense, so add an equals() method
to the LineComment class that compares the line number and message and
change CommentedFile to store a Set of unique LineComments instead of a
List. A new CommentedFile(String fileName) constructor can be used to
create an empty CommentedFile object which can then be populated using
the addLineComment() method.