Also use comment equality in CommentAdded events
When comparing two CommentAdded events to see if they're equal, we
should not only take the change and the patchset into account, but also
the comment that triggered the event. Override the hashCode and equals
methods in CommentAdded to check the comment in addition to the fields
already used by ChangeBasedEvent (its superclass).