Added a new change kind, NO_CHANGE
Since 2015, NO_CHANGE is a possible kind for a change.
This means that only the SHA-1 differs, so even more trivial
than a trivial rebase or no code change.
This should be checked by the gerrit-trigger if either exclude trivial
rebase or exclude no code change is selected, thus it needs
to be added here.